+++ packages/compiler/src/config/config-to-options.ts
@@ -1,7 +1,7 @@
 import { createDiagnosticCollector, getDirectoryPath, normalizePath } from "../core/index.js";
 import { CompilerOptions } from "../core/options.js";
 import { CompilerHost, Diagnostic } from "../core/types.js";
-import { deepClone, omitUndefined } from "../core/util.js";
+import { deepClone, doIO, omitUndefined } from "../core/util.js";
 import { expandConfigVariables } from "./config-interpolation.js";
 import { loadTypeSpecConfigForPath, validateConfigPathsAbsolute } from "./config-loader.js";
 import { EmitterOptions, TypeSpecConfig } from "./types.js";
@@ -45,10 +45,10 @@ export async function resolveCompilerOptions(
   const cwd = normalizePath(options.cwd ?? process.cwd());
   const diagnostics = createDiagnosticCollector();
 
-  const entrypointStat = await host.stat(options.entrypoint);
+  const entrypointStat = await doIO(host.stat, options.entrypoint, (diag) => diagnostics.add(diag));
   const configPath =
     options.configPath ??
-    (entrypointStat.isDirectory() ? options.entrypoint : getDirectoryPath(options.entrypoint));
+    (entrypointStat?.isDirectory() ? options.entrypoint : getDirectoryPath(options.entrypoint));
   const config = await loadTypeSpecConfigForPath(host, configPath);
   config.diagnostics.forEach((x) => diagnostics.add(x));
