+++ packages/compiler/test/config/config.test.ts
@@ -1,17 +1,14 @@
-import { deepStrictEqual, fail, strictEqual } from "assert";
-import { dirname, join, normalize, resolve } from "path";
+import { deepStrictEqual, strictEqual } from "assert";
+import { dirname, join } from "path";
 import { fileURLToPath } from "url";
 import { TypeSpecConfigJsonSchema } from "../../src/config/config-schema.js";
-import {
-  TypeSpecRawConfig,
-  loadTypeSpecConfigForPath,
-  resolveCompilerOptions,
-} from "../../src/config/index.js";
+import { TypeSpecRawConfig, loadTypeSpecConfigForPath } from "../../src/config/index.js";
 import { createSourceFile } from "../../src/core/diagnostics.js";
 import { NodeHost } from "../../src/core/node-host.js";
 import { createJSONSchemaValidator } from "../../src/core/schema-validator.js";
+import { resolvePath } from "../../src/index.js";
 
-const __scenarioRoot = resolve(
+const scenarioRoot = resolvePath(
   dirname(fileURLToPath(import.meta.url)),
   "../../../test/config/scenarios"
 );
@@ -19,7 +16,7 @@ const __scenarioRoot = resolve(
 describe("compiler: config file loading", () => {
   describe("file discovery", () => {
     const loadTestConfig = async (path: string, errorIfNotFound: boolean = true) => {
-      const fullPath = join(__scenarioRoot, path);
+      const fullPath = join(scenarioRoot, path);
       const { filename, projectRoot, ...config } = await loadTypeSpecConfigForPath(
         NodeHost,
         fullPath,
@@ -151,61 +148,6 @@ describe("compiler: config file loading", () => {
 
     it("succeeds if config is valid", () => {
       deepStrictEqual(validate({ emitters: { openapi: {} } }), []);
-    });
-  });
-});
-
-describe("compiler: resolve compiler options", () => {
-  const tspOutputPath = normalize(`${process.cwd()}/tsp-output`);
-  describe("file discovery", () => {
-    const resolveOptions = async (path: string) => {
-      const fullPath = join(__scenarioRoot, path);
-      const [options, diagnostics] = await resolveCompilerOptions(NodeHost, {
-        entrypoint: fullPath, // not really used here
-        configPath: fullPath,
-      });
-      return { options, diagnostics };
-    };
-
-    it("resolves full path to custom config file", async () => {
-      const config = await resolveOptions("custom/myConfig.yaml");
-      config.options.config = normalize(config.options.config!);
-      config.options.outputDir = normalize(config.options.outputDir!);
-
-      deepStrictEqual(config, {
-        diagnostics: [],
-        options: {
-          config: join(__scenarioRoot, "custom/myConfig.yaml"),
-          emit: ["openapi"],
-          options: {},
-          outputDir: tspOutputPath,
-        },
-      });
-    });
-
-    it("throws error when config file not found", async () => {
-      try {
-        await resolveOptions("not-found.yaml");
-      } catch (error) {
-        return;
-      }
-      fail("should throw error");
-    });
-
-    it("resolves yaml config file", async () => {
-      const config = await resolveOptions("simple");
-      config.options.config = normalize(config.options.config!);
-      config.options.outputDir = normalize(config.options.outputDir!);
-
-      deepStrictEqual(config, {
-        diagnostics: [],
-        options: {
-          config: join(__scenarioRoot, "simple/tspconfig.yaml"),
-          emit: ["openapi"],
-          options: {},
-          outputDir: tspOutputPath,
-        },
-      });
     });
   });
 });
